requirements.txt
aenum==3.1.15
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
delayed-assert==0.3.6
dill==0.3.7
exceptiongroup==1.1.3
execnet==2.0.2
Faker==19.3.1
h11==0.14.0
helium==3.2.5
idna==3.4
iniconfig==2.0.0
mysql-connector-python==8.1.0
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
protobuf==4.21.12
py==1.11.0
PySocks==1.7.1
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-reportportal==5.2.1
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
reportportal-client==5.4.0
requests==2.31.0
selenium==4.11.2
selenium-page-factory==2.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.16
webdriver-manager==4.0.0
wsproto==1.2.0
